+++ src/main/java/mx/axxib/aforedigitalgt/serv/SolicitudMatrimonioDesempleoServ.java
@@ -49,7 +49,7 @@ public class SolicitudMatrimonioDesempleoServ extends ServiceBase{
 	}
 	
 	//public List<FopagosListOut> getFopagos(Integer numSolicitud,String nss,String cuenta,String nombre) throws AforeException {
-	public FopagosListOut getFopagos(Integer numSolicitud,String nss,String cuenta,String nombre) throws AforeException {
+	public FopagosListOut getFopagos(Long numSolicitud,String nss,String cuenta,String nombre) throws AforeException {
 		try {
 			return dao.getFopagos(numSolicitud,nss,cuenta,nombre);
 		}catch (Exception e) {
